+++ funcs_test.go
@@ -22,6 +22,30 @@ func render(src string, input interface{}) (string, error) {
 	return out.String(), nil
 }
 
+type funcTestCase struct {
+	src    string
+	input  interface{}
+	expect string
+}
+
+type funcTestCases []funcTestCase
+
+func (testCases funcTestCases) Run(t *testing.T) {
+	t.Helper()
+
+	for _, testCase := range testCases {
+		actual, err := render(testCase.src, testCase.input)
+
+		if err != nil {
+			t.Errorf("unexepcted error: %v\n%#v", err, testCase)
+			continue
+		}
+		if actual != testCase.expect {
+			t.Errorf("incorrect result: %q\n%#v", actual, testCase)
+		}
+	}
+}
+
 func Test_templateFuncs_split(t *testing.T) {
 	src := `{{ $ | split ":" }}`
 	input := "quick:brown:fox:jumps"
@@ -80,4 +104,64 @@ func Test_templateFuncs_after(t *testing.T) {
 	if actual != expect {
 		t.Fatalf("incorrect: %q, want %q", actual, expect)
 	}
+}
+
+func Test_tempateFuncs_nonempty(t *testing.T) {
+	testCases := funcTestCases{
+		{
+			`{{ $ | nonempty }}`,
+			[]string{},
+			"[]",
+		},
+		{
+			`{{ $ | nonempty }}`,
+			[]string{"", ""},
+			"[]",
+		},
+		{
+			`{{ $ | nonempty }}`,
+			[]string{"quick", "", "brown", "", "fox"},
+			"[quick brown fox]",
+		},
+	}
+	testCases.Run(t)
+}
+
+func Test_tempateFuncs_strip(t *testing.T) {
+	testCases := funcTestCases{
+		// Strings
+		{
+			`{{ $ | strip }}`,
+			"",
+			"",
+		},
+		{
+			`{{ $ | strip }}`,
+			" ",
+			"",
+		},
+		{
+			`{{ $ | strip }}`,
+			"  quick brown fox  ",
+			"quick brown fox",
+		},
+		{
+			`{{ $ | strip }}`,
+			"\t\nquick brown fox\n\n",
+			"quick brown fox",
+		},
+
+		// Arrays
+		{
+			`{{ $ | strip }}`,
+			[]string{},
+			"[]",
+		},
+		{
+			`{{ $ | strip }}`,
+			[]string{" quick ", " brown ", " fox "},
+			"[quick brown fox]",
+		},
+	}
+	testCases.Run(t)
 }